fix #7357 feat(nimbus): move targeting into its own app
Because

* Targeting has fairly isolated concerns
* The targeting configurations are very high traffic for changes
* Having to explicitly configure the targeting dictionaries when making changes is annoying

This commit

* Moves targeting configs into their own app
* Auto generates the targeting config dictionaries instead of explicitly declaring them
* Changes the default 'no targeting' to an explicit slug 'no_targeting' instead of an empty string
* Migrates existing experiments from the old to new default
