+++ app/experimenter/experiments/constants.py
@@ -147,508 +147,6 @@ class Application(models.TextChoices):
     )
 
 
-@dataclass
-class NimbusTargetingConfig:
-    name: str
-    slug: str
-    description: str
-    targeting: str
-    desktop_telemetry: str
-    application_choice_names: list[str]
-
-
-TARGETING_STICKY = "experiment.slug in activeExperiments"
-TARGETING_HAS_PIN = "!doesAppNeedPin"
-TARGETING_NEED_DEFAULT = "!isDefaultBrowser"
-TARGETING_PROFILE28DAYS = "(currentDate|date - profileAgeCreated|date) / 86400000 >= 28"
-TARGETING_WIN1903 = "os.windowsBuildNumber >= 18362"
-
-TARGETING_NO_TARGETING = NimbusTargetingConfig(
-    name="No Targeting",
-    slug="",
-    description="All users",
-    targeting="",
-    desktop_telemetry="",
-    application_choice_names=[a.name for a in Application],
-)
-
-TARGETING_FIRST_RUN = NimbusTargetingConfig(
-    name="First start-up users",
-    slug="first_run",
-    description=("First start-up users (e.g. for about:welcome)"),
-    targeting=("(({is_first_startup} && {not_see_aw}) || {sticky})").format(
-        is_first_startup="isFirstStartup",
-        not_see_aw="!('trailhead.firstrun.didSeeAboutWelcome'|preferenceValue)",
-        sticky=TARGETING_STICKY,
-    ),
-    desktop_telemetry=("payload.info.profile_subsession_counter = 1"),
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_FIRST_RUN_CHROME_ATTRIBUTION = NimbusTargetingConfig(
-    name="First start-up users from Chrome",
-    slug="first_run_chrome",
-    description=(
-        "First start-up users (e.g. for about:welcome) who download Firefox "
-        "from Chrome"
-    ),
-    targeting=("{first_run} && attributionData.ua == 'chrome'").format(
-        first_run=TARGETING_FIRST_RUN.targeting
-    ),
-    desktop_telemetry=(
-        "{first_run} AND environment.settings.attribution.ua = 'chrome'"
-    ).format(first_run=TARGETING_FIRST_RUN.desktop_telemetry),
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_FIRST_RUN_WINDOWS_1903_NEWER = NimbusTargetingConfig(
-    name="First start-up users on Windows 10 1903 (build 18362) or newer",
-    slug="first_run_win1903",
-    description="First start-up users (e.g. for about:welcome) on Windows 1903+",
-    targeting=("{first_run} && os.windowsBuildNumber >= 18362").format(
-        first_run=TARGETING_FIRST_RUN.targeting
-    ),
-    desktop_telemetry=(
-        "{first_run} AND environment.system.os.windows_build_number >= 18362"
-    ).format(first_run=TARGETING_FIRST_RUN.desktop_telemetry),
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_NOT_TCP_STUDY = NimbusTargetingConfig(
-    name="Exclude users in the TCP revenue study",
-    slug="not_tcp_study",
-    description="Exclude users with certain search codes set",
-    targeting="!'browser.search.param.google_channel_us'|preferenceValue('')|regExpMatch"
-    "('^[ntc]us5$') && !'browser.search.param.google_channel_row'|preferenceValue('')|"
-    "regExpMatch('^[ntc]row5$')",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_NOT_TCP_STUDY_FIRST_RUN = NimbusTargetingConfig(
-    name="First start-up users excluding TCP revenue study",
-    slug="not_tcp_study_first_run",
-    description="First start-up users excluding certain search codes",
-    targeting="{first_run} && {not_tcp_study}".format(
-        first_run=TARGETING_FIRST_RUN.targeting,
-        not_tcp_study=TARGETING_NOT_TCP_STUDY.targeting,
-    ),
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_WINDOWS_WITH_USERCHOICE = NimbusTargetingConfig(
-    name="Users on Windows with UserChoice support",
-    slug="windows_userchoice",
-    description=(
-        "Users on Windows with UserChoice support (version 1809+/build ID 17763+)"
-    ),
-    targeting="os.windowsBuildNumber >= 17763",
-    desktop_telemetry="environment.system.os.windows_build_number >= 17763",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_WINDOWS_WITH_USERCHOICE_FIRST_RUN = NimbusTargetingConfig(
-    name="First start-up users on Windows with UserChoice support",
-    slug="windows_userchoice_first_run",
-    description=(
-        "First start-up users (e.g. for about:welcome) on Windows with "
-        "UserChoice support (version 1809+/build ID 17763+)"
-    ),
-    targeting="{first_run} && {user_choice}".format(
-        first_run=TARGETING_FIRST_RUN.targeting,
-        user_choice=TARGETING_WINDOWS_WITH_USERCHOICE.targeting,
-    ),
-    desktop_telemetry=("{first_run} AND {user_choice}").format(
-        first_run=TARGETING_FIRST_RUN.desktop_telemetry,
-        user_choice=TARGETING_WINDOWS_WITH_USERCHOICE.desktop_telemetry,
-    ),
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_FX95_DESKTOP_USERS = NimbusTargetingConfig(
-    name="Desktop Users on Fx95",
-    slug="fx95_desktop_users",
-    description=("Firefox 95 Desktop users"),
-    targeting=(
-        "(version|versionCompare('95.!') >= 0) && (version|versionCompare('96.!') < 0)"
-    ),
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_MOBILE_NEW_USER = NimbusTargetingConfig(
-    name="New Users on Mobile",
-    slug="mobile_new_users",
-    description=("New users on mobile who installed the app less than a week ago"),
-    targeting=("is_already_enrolled || days_since_install < 7"),
-    desktop_telemetry="",
-    application_choice_names=(
-        Application.FENIX.name,
-        Application.IOS.name,
-        Application.FOCUS_ANDROID.name,
-        Application.FOCUS_IOS.name,
-        Application.KLAR_ANDROID.name,
-        Application.KLAR_IOS.name,
-    ),
-)
-
-TARGETING_MOBILE_RECENTLY_UPDATED = NimbusTargetingConfig(
-    name="Recently Updated Users",
-    slug="mobile_recently_updated_users",
-    description=(
-        "Users who updated their app within the last week. "
-        "This excludes users who are new users"
-    ),
-    targeting=(
-        "is_already_enrolled || (days_since_update < 7 && days_since_install >= 7)"
-    ),
-    desktop_telemetry="",
-    application_choice_names=(
-        Application.FENIX.name,
-        Application.IOS.name,
-        Application.FOCUS_ANDROID.name,
-        Application.FOCUS_IOS.name,
-        Application.KLAR_ANDROID.name,
-        Application.KLAR_IOS.name,
-    ),
-)
-
-
-TARGETING_HOMEPAGE_GOOGLE = NimbusTargetingConfig(
-    name="Homepage set to google.com",
-    slug="homepage_google_dot_com",
-    description="Users with their Homepage set to google.com",
-    targeting=(
-        "!homePageSettings.isDefault && "
-        "homePageSettings.isCustomUrl && "
-        "homePageSettings.urls[.host == 'google.com']|length > 0"
-    ),
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_URLBAR_FIREFOX_SUGGEST = NimbusTargetingConfig(
-    name="Urlbar (Firefox Suggest)",
-    slug="urlbar_firefox_suggest",
-    description="Users with the default search suggestion showing order",
-    targeting="'browser.urlbar.showSearchSuggestionsFirst'|preferenceValue",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_MAC_ONLY = NimbusTargetingConfig(
-    name="Mac OS users only",
-    slug="mac_only",
-    description="All users with Mac OS",
-    targeting="os.isMac",
-    desktop_telemetry="environment.system.os.name = 'Darwin'",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_NO_ENTERPRISE = NimbusTargetingConfig(
-    name="No enterprise users",
-    slug="no_enterprise_users",
-    description="Exclude users with active enterpries policies",
-    targeting="!hasActiveEnterprisePolicies",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_NO_ENTERPRISE_OR_PAST_VPN = NimbusTargetingConfig(
-    name="No enterprise or past VPN use",
-    slug="no_enterprise_or_past_vpn",
-    description="Exclude users who have used Mozilla VPN or who are enterprise users",
-    targeting=(
-        f"{TARGETING_NO_ENTERPRISE.targeting} && "
-        '!("e6eb0d1e856335fc" in attachedFxAOAuthClients|mapToProperty("id"))'
-    ),
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_NO_ENTERPRISE_OR_RECENT_VPN = NimbusTargetingConfig(
-    name="No enterprise and no VPN connection in the last 30 days",
-    slug="no_enterprise_or_last_30d_vpn_use",
-    description="Exclude enterprise & users who have used MozVPN in the last 30 days",
-    targeting=(
-        f"{TARGETING_NO_ENTERPRISE.targeting} && "
-        '(("e6eb0d1e856335fc" in attachedFxAOAuthClients|mapToProperty("id")) ? '
-        '(attachedFxAOAuthClients[.id == "e6eb0d1e856335fc"].lastAccessedDaysAgo > 29) : '
-        "true)"
-    ),
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-INFREQUENT_USER_URIS = NimbusTargetingConfig(
-    name="Infrequent user (uris)",
-    slug="infrequent_user_uris",
-    description="Between 1 and 6 days of activity in the past 28 days",
-    targeting=f"{TARGETING_STICKY} || {TARGETING_PROFILE28DAYS} && "
-    "userMonthlyActivity|length >= 1 && userMonthlyActivity|length <= 6",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-INFREQUENT_USER_NEED_PIN = NimbusTargetingConfig(
-    name="Infrequent user (need pin)",
-    slug="infrequent_user_need_pin",
-    description="Between 1 and 6 days of activity in the past 28 days needing pin",
-    targeting=f"{INFREQUENT_USER_URIS.targeting} && doesAppNeedPin",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-INFREQUENT_USER_NEED_DEFAULT = NimbusTargetingConfig(
-    name="Infrequent user (need default)",
-    slug="infrequent_user_need_default",
-    description="Between 1 and 6 days of activity in the past 28 days needing default",
-    targeting=f"{INFREQUENT_USER_URIS.targeting} && {TARGETING_NEED_DEFAULT}",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-INFREQUENT_USER_NEED_DEFAULT_HAS_PIN = NimbusTargetingConfig(
-    name="Infrequent user (need default, has pin)",
-    slug="infrequent_user_need_default_has_pin",
-    description="Between 1 & 6 days activity in past 28 days need default w/ pin",
-    targeting=f"{INFREQUENT_USER_NEED_DEFAULT.targeting} && {TARGETING_HAS_PIN}",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-INFREQUENT_USER_HAS_DEFAULT_NEED_PIN = NimbusTargetingConfig(
-    name="Infrequent user (has default, need pin)",
-    slug="infrequent_user_has_default_need_pin",
-    description="Between 1 & 6 days activity in past 28 days w/ default need pin",
-    targeting=f"{INFREQUENT_USER_NEED_PIN.targeting} && isDefaultBrowser",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-INFREQUENT_WIN_USER_NEED_PIN = NimbusTargetingConfig(
-    name="Infrequent Windows user (need pin)",
-    slug="infrequent_windows_user_need_pin",
-    description="Between 1 and 6 days of activity in the past 28 days needing pin on Win",
-    targeting=f"{INFREQUENT_USER_NEED_PIN.targeting} && os.isWindows",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-INFREQUENT_WIN_USER_URIS = NimbusTargetingConfig(
-    name="Infrequent Windows 1903+ user (need default)",
-    slug="infrequent_win_user_uris",
-    description="Infrequent non default users of past 28 days, on Windows 1903+",
-    targeting=f"{INFREQUENT_USER_NEED_DEFAULT.targeting} && {TARGETING_WIN1903}",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CASUAL_USER_URIS = NimbusTargetingConfig(
-    name="Casual user (uris)",
-    slug="casual_user_uris",
-    description="Between 7 and 13 days of activity in the past 28 days",
-    targeting=f"{TARGETING_STICKY} || {TARGETING_PROFILE28DAYS} && "
-    "userMonthlyActivity|length >= 7 && userMonthlyActivity|length <= 13",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CASUAL_USER_NEED_PIN = NimbusTargetingConfig(
-    name="Casual user (need pin)",
-    slug="casual_user_need_pin",
-    description="Between 7 and 13 days of activity in the past 28 days needing pin",
-    targeting=f"{CASUAL_USER_URIS.targeting} && doesAppNeedPin",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CASUAL_USER_NEED_DEFAULT = NimbusTargetingConfig(
-    name="Casual user (need default)",
-    slug="casual_user_need_default",
-    description="Between 7 and 14 days of activity in the past 28 days needing default",
-    targeting=f"{CASUAL_USER_URIS.targeting} && {TARGETING_NEED_DEFAULT}",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CASUAL_USER_NEED_DEFAULT_HAS_PIN = NimbusTargetingConfig(
-    name="Casual user (need default, has pin)",
-    slug="casual_user_need_default_has_pin",
-    description="Between 7 and 14 days of activity in past 28 days need default w/ pin",
-    targeting=f"{CASUAL_USER_NEED_DEFAULT.targeting} && {TARGETING_HAS_PIN}",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CASUAL_USER_HAS_DEFAULT_NEED_PIN = NimbusTargetingConfig(
-    name="Casual user (has default, need pin)",
-    slug="casual_user_has_default_need_pin",
-    description="Between 7 and 14 days of activity in past 28 days w/ default need pin",
-    targeting=f"{CASUAL_USER_NEED_PIN.targeting} && isDefaultBrowser",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-REGULAR_USER_URIS = NimbusTargetingConfig(
-    name="Regular user (uris)",
-    slug="regular_user_uris",
-    description="Between 14 and 20 days of activity in the past 28 days",
-    targeting=f"{TARGETING_STICKY} || {TARGETING_PROFILE28DAYS} && "
-    "userMonthlyActivity|length >= 14 && userMonthlyActivity|length <= 20",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-REGULAR_USER_NEED_PIN = NimbusTargetingConfig(
-    name="Regular user (need pin)",
-    slug="regular_user_need_pin",
-    description="Between 14 and 20 days of activity in the past 28 days needing pin",
-    targeting=f"{REGULAR_USER_URIS.targeting} && doesAppNeedPin",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-REGULAR_USER_NEED_DEFAULT = NimbusTargetingConfig(
-    name="Regular user (need default)",
-    slug="regular_user_need_default",
-    description="Between 14 and 20 days of activity in the past 28 days needing default",
-    targeting=f"{REGULAR_USER_URIS.targeting} && {TARGETING_NEED_DEFAULT}",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-REGULAR_USER_NEED_DEFAULT_HAS_PIN = NimbusTargetingConfig(
-    name="Regular user (need default, has pin)",
-    slug="regular_user_need_default_has_pin",
-    description="Between 14 and 20 days of activity in past 28 days need default w/ pin",
-    targeting=f"{REGULAR_USER_NEED_DEFAULT.targeting} && {TARGETING_HAS_PIN}",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-REGULAR_USER_HAS_DEFAULT_NEED_PIN = NimbusTargetingConfig(
-    name="Regular user (has default, need pin)",
-    slug="regular_user_has_default_need_pin",
-    description="Between 14 and 20 days of activity in past 28 days w/ default need pin",
-    targeting=f"{REGULAR_USER_NEED_PIN.targeting} && isDefaultBrowser",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-REGULAR_USER_USES_FXA = NimbusTargetingConfig(
-    name="Regular user (uses Firefox Accounts)",
-    slug="regular_user_uses_fxa",
-    description="Between 14 and 20 days of activity in the past 28 days signed in to FxA",
-    targeting=f"{REGULAR_USER_URIS.targeting} && isFxAEnabled && usesFirefoxSync",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CORE_USER_URIS = NimbusTargetingConfig(
-    name="Core user (uris)",
-    slug="core_user_uris",
-    description="At least 21 days of activity in the past 28 days",
-    targeting=f"{TARGETING_STICKY} || {TARGETING_PROFILE28DAYS} && "
-    "userMonthlyActivity|length >= 21",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CORE_USER_NEED_PIN = NimbusTargetingConfig(
-    name="Core user (need pin)",
-    slug="core_user_need_pin",
-    description="At least 21 days of activity in the past 28 days needing pin",
-    targeting=f"{CORE_USER_URIS.targeting} && doesAppNeedPin",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CORE_USER_NEED_DEFAULT = NimbusTargetingConfig(
-    name="Core user (need default)",
-    slug="core_user_need_default",
-    description="At least 21 days of activity in the past 28 days needing default",
-    targeting=f"{CORE_USER_URIS.targeting} && {TARGETING_NEED_DEFAULT}",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CORE_USER_NEED_DEFAULT_HAS_PIN = NimbusTargetingConfig(
-    name="Core user (need default, has pin)",
-    slug="core_user_need_default_has_pin",
-    description="At least 21 days of activity in past 28 days need default w/ pin",
-    targeting=f"{CORE_USER_NEED_DEFAULT.targeting} && {TARGETING_HAS_PIN}",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-CORE_USER_HAS_DEFAULT_NEED_PIN = NimbusTargetingConfig(
-    name="Core user (has default, need pin)",
-    slug="core_user_has_default_need_pin",
-    description="At least 21 days of activity in past 28 days w/ default need pin",
-    targeting=f"{CORE_USER_NEED_PIN.targeting} && isDefaultBrowser",
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_POCKET_COMMON = NimbusTargetingConfig(
-    name="Pocket Common Filters",
-    slug="pocket_common",
-    description="All newtab sections are default",
-    targeting="""
-    (
-        (
-            (
-                !('browser.newtabpage.enabled'|preferenceIsUserSet)
-                ||
-                !('browser.startup.homepage'|preferenceIsUserSet)
-            )
-            &&
-            !('browser.newtabpage.activity-stream.showSearch'|preferenceIsUserSet)
-            &&
-            !('browser.newtabpage.activity-stream.feeds.topsites'|preferenceIsUserSet)
-            &&
-            !('browser.newtabpage.activity-stream.feeds.section.topstories'|preferenceIsUserSet)
-            &&
-            !('browser.newtabpage.activity-stream.feeds.section.highlights'|preferenceIsUserSet)
-        )
-    )
-    """,
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_PIP_NEVER_USED = NimbusTargetingConfig(
-    name="PiP Never Used",
-    slug="pip_never_used",
-    description="Users that have never used Picture in Picture",
-    targeting="""
-    (
-        !('media.videocontrols.picture-in-picture.video-toggle.has-used'|preferenceValue)
-        &&
-        ('media.videocontrols.picture-in-picture.enabled'|preferenceValue)
-        &&
-        ('media.videocontrols.picture-in-picture.video-toggle.enabled'|preferenceValue)
-    )
-    """,
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-TARGETING_PIP_NEVER_USED_STICKY = NimbusTargetingConfig(
-    name="PiP Never Used (Sticky)",
-    slug="pip_never_used_sticky",
-    description="Users that have never used Picture in Picture, with sticky enrollment",
-    targeting="(({pip}) || ({sticky}))".format(
-        pip=TARGETING_PIP_NEVER_USED.targeting,
-        sticky=TARGETING_STICKY,
-    ),
-    desktop_telemetry="",
-    application_choice_names=(Application.DESKTOP.name,),
-)
-
-
 class NimbusConstants(object):
     class Status(models.TextChoices):
         DRAFT = "Draft"
@@ -852,231 +350,6 @@ class NimbusConstants(object):
 
     EMAIL_EXPERIMENT_END_SUBJECT = "Action required: Please turn off your Experiment"
     EMAIL_ENROLLMENT_END_SUBJECT = "Action required: Please end experiment enrollment"
-
-    TARGETING_VERSION = "version|versionCompare('{version}') >= 0"
-    TARGETING_CHANNEL = 'browserSettings.update.channel == "{channel}"'
-
-    TARGETING_CONFIGS = {
-        TARGETING_NO_TARGETING.slug: TARGETING_NO_TARGETING,
-        TARGETING_FIRST_RUN.slug: TARGETING_FIRST_RUN,
-        TARGETING_FIRST_RUN_CHROME_ATTRIBUTION.slug: (
-            TARGETING_FIRST_RUN_CHROME_ATTRIBUTION
-        ),
-        TARGETING_FIRST_RUN_WINDOWS_1903_NEWER.slug: (
-            TARGETING_FIRST_RUN_WINDOWS_1903_NEWER
-        ),
-        TARGETING_HOMEPAGE_GOOGLE.slug: TARGETING_HOMEPAGE_GOOGLE,
-        TARGETING_URLBAR_FIREFOX_SUGGEST.slug: TARGETING_URLBAR_FIREFOX_SUGGEST,
-        TARGETING_MAC_ONLY.slug: TARGETING_MAC_ONLY,
-        TARGETING_NO_ENTERPRISE.slug: TARGETING_NO_ENTERPRISE,
-        TARGETING_FX95_DESKTOP_USERS.slug: TARGETING_FX95_DESKTOP_USERS,
-        TARGETING_NOT_TCP_STUDY.slug: TARGETING_NOT_TCP_STUDY,
-        TARGETING_NOT_TCP_STUDY_FIRST_RUN.slug: TARGETING_NOT_TCP_STUDY_FIRST_RUN,
-        TARGETING_WINDOWS_WITH_USERCHOICE.slug: TARGETING_WINDOWS_WITH_USERCHOICE,
-        TARGETING_WINDOWS_WITH_USERCHOICE_FIRST_RUN.slug: (
-            TARGETING_WINDOWS_WITH_USERCHOICE_FIRST_RUN
-        ),
-        TARGETING_MOBILE_NEW_USER.slug: TARGETING_MOBILE_NEW_USER,
-        TARGETING_MOBILE_RECENTLY_UPDATED.slug: TARGETING_MOBILE_RECENTLY_UPDATED,
-        TARGETING_NO_ENTERPRISE_OR_PAST_VPN.slug: TARGETING_NO_ENTERPRISE_OR_PAST_VPN,
-        TARGETING_NO_ENTERPRISE_OR_RECENT_VPN.slug: TARGETING_NO_ENTERPRISE_OR_RECENT_VPN,
-        INFREQUENT_USER_URIS.slug: INFREQUENT_USER_URIS,
-        INFREQUENT_USER_NEED_PIN.slug: INFREQUENT_USER_NEED_PIN,
-        INFREQUENT_USER_NEED_DEFAULT.slug: INFREQUENT_USER_NEED_DEFAULT,
-        INFREQUENT_USER_NEED_DEFAULT_HAS_PIN.slug: INFREQUENT_USER_NEED_DEFAULT_HAS_PIN,
-        INFREQUENT_USER_HAS_DEFAULT_NEED_PIN.slug: INFREQUENT_USER_HAS_DEFAULT_NEED_PIN,
-        INFREQUENT_WIN_USER_NEED_PIN.slug: INFREQUENT_WIN_USER_NEED_PIN,
-        INFREQUENT_WIN_USER_URIS.slug: INFREQUENT_WIN_USER_URIS,
-        CASUAL_USER_URIS.slug: CASUAL_USER_URIS,
-        CASUAL_USER_NEED_PIN.slug: CASUAL_USER_NEED_PIN,
-        CASUAL_USER_NEED_DEFAULT.slug: CASUAL_USER_NEED_DEFAULT,
-        CASUAL_USER_NEED_DEFAULT_HAS_PIN.slug: CASUAL_USER_NEED_DEFAULT_HAS_PIN,
-        CASUAL_USER_HAS_DEFAULT_NEED_PIN.slug: CASUAL_USER_HAS_DEFAULT_NEED_PIN,
-        REGULAR_USER_URIS.slug: REGULAR_USER_URIS,
-        REGULAR_USER_NEED_PIN.slug: REGULAR_USER_NEED_PIN,
-        REGULAR_USER_NEED_DEFAULT.slug: REGULAR_USER_NEED_DEFAULT,
-        REGULAR_USER_NEED_DEFAULT_HAS_PIN.slug: REGULAR_USER_NEED_DEFAULT_HAS_PIN,
-        REGULAR_USER_HAS_DEFAULT_NEED_PIN.slug: REGULAR_USER_HAS_DEFAULT_NEED_PIN,
-        REGULAR_USER_USES_FXA.slug: REGULAR_USER_USES_FXA,
-        CORE_USER_URIS.slug: CORE_USER_URIS,
-        CORE_USER_NEED_PIN.slug: CORE_USER_NEED_PIN,
-        CORE_USER_NEED_DEFAULT.slug: CORE_USER_NEED_DEFAULT,
-        CORE_USER_NEED_DEFAULT_HAS_PIN.slug: CORE_USER_NEED_DEFAULT_HAS_PIN,
-        CORE_USER_HAS_DEFAULT_NEED_PIN.slug: CORE_USER_HAS_DEFAULT_NEED_PIN,
-        TARGETING_POCKET_COMMON.slug: TARGETING_POCKET_COMMON,
-        TARGETING_PIP_NEVER_USED.slug: TARGETING_PIP_NEVER_USED,
-        TARGETING_PIP_NEVER_USED_STICKY.slug: TARGETING_PIP_NEVER_USED_STICKY,
-    }
-
-    class TargetingConfig(models.TextChoices):
-        NO_TARGETING = TARGETING_NO_TARGETING.slug, TARGETING_NO_TARGETING.name
-        TARGETING_FIRST_RUN = TARGETING_FIRST_RUN.slug, TARGETING_FIRST_RUN.name
-        TARGETING_FIRST_RUN_CHROME_ATTRIBUTION = (
-            TARGETING_FIRST_RUN_CHROME_ATTRIBUTION.slug
-        ), TARGETING_FIRST_RUN_CHROME_ATTRIBUTION.name
-        TARGETING_FIRST_RUN_WINDOWS_1903_NEWER = (
-            TARGETING_FIRST_RUN_WINDOWS_1903_NEWER.slug
-        ), TARGETING_FIRST_RUN_WINDOWS_1903_NEWER.name
-        TARGETING_HOMEPAGE_GOOGLE = (
-            TARGETING_HOMEPAGE_GOOGLE.slug,
-            TARGETING_HOMEPAGE_GOOGLE.name,
-        )
-        TARGETING_URLBAR_FIREFOX_SUGGEST = (
-            TARGETING_URLBAR_FIREFOX_SUGGEST.slug,
-            TARGETING_URLBAR_FIREFOX_SUGGEST.name,
-        )
-        TARGETING_MAC_ONLY = (TARGETING_MAC_ONLY.slug, TARGETING_MAC_ONLY.name)
-        TARGETING_NO_ENTERPRISE = (
-            TARGETING_NO_ENTERPRISE.slug,
-            TARGETING_NO_ENTERPRISE.name,
-        )
-        TARGETING_FX95_DESKTOP_USERS = (
-            TARGETING_FX95_DESKTOP_USERS.slug,
-            TARGETING_FX95_DESKTOP_USERS.name,
-        )
-        TARGETING_NOT_TCP_STUDY = (
-            TARGETING_NOT_TCP_STUDY.slug,
-            TARGETING_NOT_TCP_STUDY.name,
-        )
-        TARGETING_NOT_TCP_STUDY_FIRST_RUN = (
-            TARGETING_NOT_TCP_STUDY_FIRST_RUN.slug,
-            TARGETING_NOT_TCP_STUDY_FIRST_RUN.name,
-        )
-        TARGETING_WINDOWS_WITH_USERCHOICE = (
-            TARGETING_WINDOWS_WITH_USERCHOICE.slug,
-            TARGETING_WINDOWS_WITH_USERCHOICE.name,
-        )
-        TARGETING_WINDOWS_WITH_USERCHOICE_FIRST_RUN = (
-            TARGETING_WINDOWS_WITH_USERCHOICE_FIRST_RUN.slug,
-            TARGETING_WINDOWS_WITH_USERCHOICE_FIRST_RUN.name,
-        )
-        TARGETING_MOBILE_NEW_USER = (
-            TARGETING_MOBILE_NEW_USER.slug,
-            TARGETING_MOBILE_NEW_USER.name,
-        )
-        TARGETING_MOBILE_RECENTLY_UPDATED = (
-            TARGETING_MOBILE_RECENTLY_UPDATED.slug,
-            TARGETING_MOBILE_RECENTLY_UPDATED.name,
-        )
-        INFREQUENT_USER_URIS = (
-            INFREQUENT_USER_URIS.slug,
-            INFREQUENT_USER_URIS.name,
-        )
-        INFREQUENT_USER_NEED_PIN = (
-            INFREQUENT_USER_NEED_PIN.slug,
-            INFREQUENT_USER_NEED_PIN.name,
-        )
-        INFREQUENT_USER_NEED_DEFAULT = (
-            INFREQUENT_USER_NEED_DEFAULT.slug,
-            INFREQUENT_USER_NEED_DEFAULT.name,
-        )
-        INFREQUENT_WIN_USER_NEED_PIN = (
-            INFREQUENT_WIN_USER_NEED_PIN.slug,
-            INFREQUENT_WIN_USER_NEED_PIN.name,
-        )
-        INFREQUENT_USER_NEED_DEFAULT_HAS_PIN = (
-            INFREQUENT_USER_NEED_DEFAULT_HAS_PIN.slug,
-            INFREQUENT_USER_NEED_DEFAULT_HAS_PIN.name,
-        )
-        INFREQUENT_USER_HAS_DEFAULT_NEED_PIN = (
-            INFREQUENT_USER_HAS_DEFAULT_NEED_PIN.slug,
-            INFREQUENT_USER_HAS_DEFAULT_NEED_PIN.name,
-        )
-        INFREQUENT_WIN_USER_URIS = (
-            INFREQUENT_WIN_USER_URIS.slug,
-            INFREQUENT_WIN_USER_URIS.name,
-        )
-        CASUAL_USER_URIS = (
-            CASUAL_USER_URIS.slug,
-            CASUAL_USER_URIS.name,
-        )
-        CASUAL_USER_NEED_PIN = (
-            CASUAL_USER_NEED_PIN.slug,
-            CASUAL_USER_NEED_PIN.name,
-        )
-        CASUAL_USER_NEED_DEFAULT = (
-            CASUAL_USER_NEED_DEFAULT.slug,
-            CASUAL_USER_NEED_DEFAULT.name,
-        )
-        CASUAL_USER_NEED_DEFAULT_HAS_PIN = (
-            CASUAL_USER_NEED_DEFAULT_HAS_PIN.slug,
-            CASUAL_USER_NEED_DEFAULT_HAS_PIN.name,
-        )
-        CASUAL_USER_HAS_DEFAULT_NEED_PIN = (
-            CASUAL_USER_HAS_DEFAULT_NEED_PIN.slug,
-            CASUAL_USER_HAS_DEFAULT_NEED_PIN.name,
-        )
-        REGULAR_USER_URIS = (
-            REGULAR_USER_URIS.slug,
-            REGULAR_USER_URIS.name,
-        )
-        REGULAR_USER_NEED_PIN = (
-            REGULAR_USER_NEED_PIN.slug,
-            REGULAR_USER_NEED_PIN.name,
-        )
-        REGULAR_USER_NEED_DEFAULT = (
-            REGULAR_USER_NEED_DEFAULT.slug,
-            REGULAR_USER_NEED_DEFAULT.name,
-        )
-        REGULAR_USER_NEED_DEFAULT_HAS_PIN = (
-            REGULAR_USER_NEED_DEFAULT_HAS_PIN.slug,
-            REGULAR_USER_NEED_DEFAULT_HAS_PIN.name,
-        )
-        REGULAR_USER_HAS_DEFAULT_NEED_PIN = (
-            REGULAR_USER_HAS_DEFAULT_NEED_PIN.slug,
-            REGULAR_USER_HAS_DEFAULT_NEED_PIN.name,
-        )
-        REGULAR_USER_USES_FXA = (
-            REGULAR_USER_USES_FXA.slug,
-            REGULAR_USER_USES_FXA.name,
-        )
-        CORE_USER_URIS = (
-            CORE_USER_URIS.slug,
-            CORE_USER_URIS.name,
-        )
-        CORE_USER_NEED_PIN = (
-            CORE_USER_NEED_PIN.slug,
-            CORE_USER_NEED_PIN.name,
-        )
-        CORE_USER_NEED_DEFAULT = (
-            CORE_USER_NEED_DEFAULT.slug,
-            CORE_USER_NEED_DEFAULT.name,
-        )
-        CORE_USER_NEED_DEFAULT_HAS_PIN = (
-            CORE_USER_NEED_DEFAULT_HAS_PIN.slug,
-            CORE_USER_NEED_DEFAULT_HAS_PIN.name,
-        )
-        CORE_USER_HAS_DEFAULT_NEED_PIN = (
-            CORE_USER_HAS_DEFAULT_NEED_PIN.slug,
-            CORE_USER_HAS_DEFAULT_NEED_PIN.name,
-        )
-        TARGETING_NO_ENTERPRISE_OR_PAST_VPN = (
-            TARGETING_NO_ENTERPRISE_OR_PAST_VPN.slug,
-            TARGETING_NO_ENTERPRISE_OR_PAST_VPN.name,
-        )
-        TARGETING_NO_ENTERPRISE_OR_RECENT_VPN = (
-            TARGETING_NO_ENTERPRISE_OR_RECENT_VPN.slug,
-            TARGETING_NO_ENTERPRISE_OR_RECENT_VPN.name,
-        )
-        TARGETING_POCKET_COMMON = (
-            TARGETING_POCKET_COMMON.slug,
-            TARGETING_POCKET_COMMON.name,
-        )
-        TARGETING_PIP_NEVER_USED = (
-            TARGETING_PIP_NEVER_USED.slug,
-            TARGETING_PIP_NEVER_USED.name,
-        )
-        TARGETING_PIP_NEVER_USED_STICKY = (
-            TARGETING_PIP_NEVER_USED_STICKY.slug,
-            TARGETING_PIP_NEVER_USED_STICKY.name,
-        )
-
-    TARGETING_APPLICATION_SUPPORTED_VERSION = {
-        Application.FENIX: Version.FIREFOX_98,
-        Application.FOCUS_ANDROID: Version.FIREFOX_98,
-        Application.IOS: Version.FIREFOX_98,
-        Application.FOCUS_IOS: Version.FIREFOX_97,
-    }
 
     LANGUAGES_APPLICATION_SUPPORTED_VERSION = {
         Application.FENIX: Version.FIREFOX_102,
