+++ app/experimenter/experiments/tests/test_models.py
@@ -213,7 +213,7 @@ class TestNimbusExperiment(TestCase):
         experiment = NimbusExperimentFactory.create(
             firefox_min_version=NimbusExperiment.Version.FIREFOX_83,
             firefox_max_version=NimbusExperiment.Version.FIREFOX_95,
-            targeting_config_slug=NimbusExperiment.TargetingConfig.TARGETING_MAC_ONLY,
+            targeting_config_slug=NimbusExperiment.TargetingConfig.MAC_ONLY,
             application=NimbusExperiment.Application.DESKTOP,
             channel=NimbusExperiment.Channel.NO_CHANNEL,
             locales=[],
@@ -408,7 +408,7 @@ class TestNimbusExperiment(TestCase):
             NimbusExperimentFactory.Lifecycles.LAUNCH_APPROVE_APPROVE,
             firefox_min_version=NimbusExperiment.Version.NO_VERSION,
             firefox_max_version=NimbusExperiment.Version.FIREFOX_95,
-            targeting_config_slug=NimbusExperiment.TargetingConfig.TARGETING_MAC_ONLY,
+            targeting_config_slug=NimbusExperiment.TargetingConfig.MAC_ONLY,
             application=NimbusExperiment.Application.DESKTOP,
             channel=NimbusExperiment.Channel.NIGHTLY,
             locales=[],
@@ -434,7 +434,7 @@ class TestNimbusExperiment(TestCase):
             NimbusExperimentFactory.Lifecycles.LAUNCH_APPROVE_APPROVE,
             firefox_min_version=NimbusExperiment.Version.FIREFOX_83,
             firefox_max_version=NimbusExperiment.Version.NO_VERSION,
-            targeting_config_slug=NimbusExperiment.TargetingConfig.TARGETING_MAC_ONLY,
+            targeting_config_slug=NimbusExperiment.TargetingConfig.MAC_ONLY,
             application=NimbusExperiment.Application.DESKTOP,
             channel=NimbusExperiment.Channel.NIGHTLY,
             locales=[],
@@ -458,7 +458,7 @@ class TestNimbusExperiment(TestCase):
             NimbusExperimentFactory.Lifecycles.LAUNCH_APPROVE_APPROVE,
             firefox_min_version=NimbusExperiment.Version.NO_VERSION,
             firefox_max_version=NimbusExperiment.Version.NO_VERSION,
-            targeting_config_slug=NimbusExperiment.TargetingConfig.TARGETING_MAC_ONLY,
+            targeting_config_slug=NimbusExperiment.TargetingConfig.MAC_ONLY,
             application=NimbusExperiment.Application.DESKTOP,
             channel=NimbusExperiment.Channel.NO_CHANNEL,
             locales=[],
@@ -479,7 +479,7 @@ class TestNimbusExperiment(TestCase):
             application=NimbusExperiment.Application.DESKTOP,
             firefox_min_version=NimbusExperiment.Version.NO_VERSION,
             firefox_max_version=NimbusExperiment.Version.NO_VERSION,
-            targeting_config_slug=NimbusExperiment.TargetingConfig.TARGETING_MAC_ONLY,
+            targeting_config_slug=NimbusExperiment.TargetingConfig.MAC_ONLY,
             channel=NimbusExperiment.Channel.NO_CHANNEL,
             locales=[locale_ca, locale_us],
             countries=[],
@@ -503,7 +503,7 @@ class TestNimbusExperiment(TestCase):
             application=NimbusExperiment.Application.DESKTOP,
             firefox_min_version=NimbusExperiment.Version.NO_VERSION,
             firefox_max_version=NimbusExperiment.Version.NO_VERSION,
-            targeting_config_slug=NimbusExperiment.TargetingConfig.TARGETING_MAC_ONLY,
+            targeting_config_slug=NimbusExperiment.TargetingConfig.MAC_ONLY,
             channel=NimbusExperiment.Channel.NO_CHANNEL,
             locales=[],
             countries=[country_ca, country_us],
@@ -529,7 +529,7 @@ class TestNimbusExperiment(TestCase):
             application=NimbusExperiment.Application.DESKTOP,
             firefox_min_version=NimbusExperiment.Version.NO_VERSION,
             firefox_max_version=NimbusExperiment.Version.NO_VERSION,
-            targeting_config_slug=NimbusExperiment.TargetingConfig.TARGETING_MAC_ONLY,
+            targeting_config_slug=NimbusExperiment.TargetingConfig.MAC_ONLY,
             channel=NimbusExperiment.Channel.NO_CHANNEL,
             locales=[locale_ca, locale_us],
             countries=[country_ca, country_us],
@@ -556,7 +556,7 @@ class TestNimbusExperiment(TestCase):
             application=NimbusExperiment.Application.FENIX,
             firefox_min_version=NimbusExperiment.Version.NO_VERSION,
             firefox_max_version=NimbusExperiment.Version.NO_VERSION,
-            targeting_config_slug=targeting_config.TARGETING_MOBILE_NEW_USER,
+            targeting_config_slug=targeting_config.MOBILE_NEW_USERS,
             channel=NimbusExperiment.Channel.NO_CHANNEL,
             languages=[language_en, language_es, language_fr],
         )
@@ -585,7 +585,7 @@ class TestNimbusExperiment(TestCase):
             application=NimbusExperiment.Application.FENIX,
             firefox_min_version=NimbusExperiment.Version.NO_VERSION,
             firefox_max_version=NimbusExperiment.Version.NO_VERSION,
-            targeting_config_slug=targeting_config.TARGETING_MOBILE_NEW_USER,
+            targeting_config_slug=targeting_config.MOBILE_NEW_USERS,
             channel=NimbusExperiment.Channel.NO_CHANNEL,
             locales=[locale_ca, locale_ro, locale_de, locale_es, locale_us],
             languages=[language_en, language_es, language_fr],
