+++ app/experimenter/targeting/tests/test_targeting_configs.py
@@ -2,20 +2,20 @@ from django.test import TestCase
 from parameterized import parameterized
 from parsimonious.exceptions import ParseError
 
-from experimenter.experiments.constants import NimbusConstants
 from experimenter.experiments.tests import JEXLParser
+from experimenter.targeting.constants import TargetingConstants
 
 
 class TestTargetingConfigs(TestCase):
     def test_all_targeting_configs_defined_in_constants(self):
         self.assertEqual(
-            set([t.value for t in NimbusConstants.TargetingConfig]),
-            set(t for t in NimbusConstants.TARGETING_CONFIGS.keys()),
-            "Targeting Configs must be defined in both NimbusConstants.TargetingConfig "
-            "and NimbusConstants.TARGETING_CONFIGS",
+            set([t.value for t in TargetingConstants.TargetingConfig]),
+            set(t for t in TargetingConstants.TARGETING_CONFIGS.keys()),
+            "Targeting Configs must be defined in both "
+            "TargetingConstants.TargetingConfig and TargetingConstants.TARGETING_CONFIGS",
         )
 
-    @parameterized.expand([(t,) for t in NimbusConstants.TARGETING_CONFIGS.values()])
+    @parameterized.expand([(t,) for t in TargetingConstants.TARGETING_CONFIGS.values()])
     def test_targeting_config_has_valid_jexl(self, targeting_config):
         if targeting_config.targeting:
             try:
