+++ app/tests/integration/nimbus/conftest.py
@@ -152,7 +152,7 @@ def default_data(request):
         audience=BaseExperimentAudienceDataClass(
             channel=BaseExperimentAudienceChannels.RELEASE,
             min_version=80,
-            targeting="",
+            targeting="no_targeting",
             percentage=50.0,
             expected_clients=50,
             locale=None,
